Add callbacks for CursorAdapter
CursorLoaders were not working with our OrmLiteCursorAdapters. The
reason is that the data management system, typically the
ContentProvider, is responsible for notifying the loader when the data
has changed.

In order to get a similar result, I'm trying to maintain a list of weak
reference to Loaders created for AndroidBaseDaoImpls. This may not be
the long term solution, but it does work for the time being.